new_folder/requirements.txt
attrs
autopep8
beautifulsoup4
cachetools
certifi
chardet
click
colorama
decorator
flake8
future
geocoder
geographiclib
geopy
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
googleapis-common-protos
httplib2
idna
isort
lazy-object-proxy
lxml
mccabe
more-itertools
numpy
oauthlib
Pillow
pprintpp
protobuf
psutil
pyasn1
pyasn1-modules
pycodestyle
pyflakes
pyjokes
pylint
python-dotenv
pytz
pywhatkit # Kept for web searches/YouTube, though it uses webbrowser
ratelim
requests
requests-oauthlib
rsa
selenium # Kept for browser automation
six
soupsieve
toml
uritemplate
urllib3
wikipedia
wincertstore # Kept - Windows certificate store, not directly GUI
wrapt
# xml-python # Removed - less common than lxml
xmltodict
RealtimeTTS[all]
RealtimeSTT[all]
google.generativeai